% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_perm.R
\name{best_perm}
\alias{best_perm}
\title{Find the optimal permutation of latent classes}
\usage{
best_perm(Utrue, Uprop)
}
\arguments{
\item{Utrue}{Vector of true class labels.}

\item{Uprop}{Vector of estimated labels.}
}
\value{
A list containing the contingency table and the optimal permutation.
}
\description{
Find the optimal permutation of latent classes
}
