% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_hyb_sbm_dyn_dec.R
\name{est_pred_sbm_dyn_decoding}
\alias{est_pred_sbm_dyn_decoding}
\title{Estimation of the Hybrid DSBM model via the CEM algorithm}
\usage{
est_pred_sbm_dyn_decoding(
  Y,
  k,
  start = 0,
  tol = 10^-10,
  maxit = 100,
  Tau = NULL
)
}
\arguments{
\item{Y}{3D binary array (n x n x TT) representing the temporal adjacency matrix.}

\item{k}{Number of latent blocks/classes.}

\item{start}{Initialization strategy (0 = K-means, 1 = random, 2 = custom Tau).}

\item{tol}{Convergence tolerance for the log-likelihood.}

\item{maxit}{Maximum number of CEM iterations.}

\item{Tau}{Initial membership array (required if start = 2).}
}
\value{
List containing parameter estimates, likelihood, and classifications (cl and clv).
}
\description{
Estimation of the Hybrid DSBM model via the CEM algorithm
}
