% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdsbm-package.R
\docType{package}
\name{hdsbm-package}
\alias{hdsbm}
\alias{hdsbm-package}
\title{hdsbm: Hybrid Dynamic Stochastic Block Models}
\description{
Estimation and simulation routines for discrete-time Dynamic Stochastic Block Models via Classification EM and Viterbi decoding.
}
\author{
\strong{Maintainer}: Gianfilippo Tito \email{gianfilippotito@gmail.com}

Authors:
\itemize{
  \item Gianfilippo Tito \email{gianfilippotito@gmail.com}
}

}
\keyword{internal}
