% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_hdsbm}
\alias{toy_hdsbm}
\title{Synthetic dataset for Dynamic SBM models}
\format{
A list with 3 elements:
\describe{
\item{Y}{Three-dimensional binary array (25 x 25 x 4).}
\item{U_true}{25 x 4 matrix of true community memberships.}
\item{params}{Dimensional parameters (n, k, TT).}
}
}
\source{
Data simulated via \code{draw_sn_dyn}.
}
\usage{
toy_hdsbm
}
\description{
A sample dataset containing a dynamic network simulated under
the Hybrid DSBM framework.
}
\examples{
data(toy_hdsbm)
dim(toy_hdsbm$Y)
}
\keyword{datasets}
